Factor out Injectable from Prismable
I realized that I don’t know how to `pcoerce` the type `[a]` from an AValue, but
there's a pretty natural injection. Well, it's not really an injective because
Objects could also get in there. Not sure what to call this. `Representable`?

Anyway, it also lets me define `Injectable (Set AValue) a` which just creates a
singleton object.