Finally figured out flow sens.
So, flow sensitivity was wrong in the initial paper, and then I fixed it
and it was subtly wrong again. I finally fixed it so that flow
sensitivity accurately represents "one set of facts per program point".

The issue now is the examples. Flow sensitivity is only helpful if the
facts at one program point are different in an interesting way than from
another program point. In imperative languages this happens through
strong update. In our little functional language, information (the
store) is growing strictly monotonically. This means that a
flow-insensitive and flow-sensitive analysis are provably the same
analysis for our language. Ouch.

So there are two options: move to an imperative language, or do
something interesting with the store that causes it to grow
non-monotonically. As a chance of fate, we already have such a thing
implemented: abstract garbage collection. Abstract GC will wipe away old
references to a binding site so you get a fresh allocation the next time
you encouter it. This means that, when abstract GC is enabled, |let x :=
f 1 in let y := f 2 in ...| will not confuse the bindings of |f 1| and
|f 2| in a flow sensitive analysis, but they will still be confused in a
flow insensitive analysis.

The plan going forward is to use abstract garbage collection in the
example of flow-sensitivity since it will give different results than
both flow-insensitive and path-sensitive. This is also nice because we
introduce, define and discuss abstract gc at length in the paper, so
this will only serve to supplement the rest of the development.